python_service/requirements.txt
tensorflow==2.15.0
Pillow==10.3.0
numpy==1.26.4
fastapi==0.103.0
psycopg2-binary==2.9.9
uvicorn==0.23.2